Acrescentado SwalFire

+++ DesafioWebSite/src/app/routes/cadastro/cadastro.component.ts
@@ -3,6 +3,9 @@ import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { CadastroService } from 'src/app/services/cadastro.service';
 import {Employee} from 'src/app/models/Employee'
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
+import Swal from "sweetalert2";
+import { Router } from '@angular/router';
+import { formatDate } from '@angular/common';
 
 @Component({
   selector: 'app-cadastro',
@@ -14,13 +17,16 @@ export class CadastroComponent implements OnInit {
   form: FormGroup;
   formEdit: FormGroup;
 
+  employeeCreate = false;
+  employeeEdit = false;
+
   employees: Array<Employee> = new Array<Employee>();
 
   employeeId: string = null;
 
   @ViewChild('modalEditEmployee', { static: false }) modalEditEmployee: TemplateRef<any>;
   modalEditEmployeeRef: BsModalRef;
-  constructor(private cadastroService:CadastroService, private modalService: BsModalService) { }
+  constructor(private cadastroService:CadastroService, private modalService: BsModalService,private router: Router)  { }
 
   ngOnInit(): void {
     this.getEmployee();
@@ -55,7 +61,6 @@ export class CadastroComponent implements OnInit {
     this.cadastroService.get().subscribe(
       (response) => {
         Object.assign(this.employees, response);
-        console.log(this.employees);
       },
       (error) => {
         console.log(error);
@@ -64,26 +69,49 @@ export class CadastroComponent implements OnInit {
   }
 
   createEmployee() {
+    debugger
     let employeeForm = this.form.value as Employee;
     this.cadastroService.save(employeeForm).subscribe(
       (response) => {
-        console.log(response);
+        Swal.fire('Success', 'Colaborador Cadastrado com Sucesso', 'success');
       },
       (error) => {
-        console.log(error);
+        Swal.fire('Error', 'Não foi possível Cadastrar o Colaborador', 'error');
       }
     );
   }
 
-  deleteEmployee(employee: Employee) {
-    this.cadastroService.delete(employee._id).subscribe(
-      (response) => {
-        console.log(response);
-      },
-      (error) => {
-        console.log(error);
+  deleteEmployee(employee: Employee) 
+  {
+    Swal.fire({
+      text: `Deseja mesmo Deletar o colaborador  ${employee.nome}?`,
+      icon:"warning",
+      confirmButtonColor: "#f05050",
+      confirmButtonText: "Deletar Colaborador",
+      cancelButtonColor: "Cancelar",
+      iconHtml:'<i class="fas fa-trash"',
+      iconColor: "#f05050",
+      showCancelButton: true,
+    }).then((result) =>
+    {
+      if(result.value)
+      {
+        this.cadastroService.delete(employee._id).subscribe(
+          (response) => {
+            Swal.fire('Success', 'Colaborador Excluido com Sucesso', 'success');
+            this.router.navigate(['/']);
+            
+          },
+          (error) => {
+            Swal.fire('Error', 'Não foi possível Excluir o Colaborador', 'error');
+          }
+        );
+        this.router.navigate(['/']);
       }
-    );
+    });
+    
+    
+    
   }
 
   editEmployee() {
@@ -92,10 +120,10 @@ export class CadastroComponent implements OnInit {
     debugger
     this.cadastroService.edit(employee).subscribe(
       (response) => {
-        console.log(response);
+        Swal.fire('Success', 'Colaborador Editado com Sucesso', 'success');
       },
       (error) => {
-        console.log(error);
+        Swal.fire('Error', 'Não foi possível editar o Colaborador', 'error');
       }
     );
   }
